refactor: bindings to actions and store
